+++ super_editor/lib/src/default_editor/document_ime/supereditor_ime_interactor.dart
@@ -669,7 +669,9 @@ void insertNewLine(SuperEditorContext context) {
   if (CurrentPlatform.isWeb) {
     return;
   }
-  context.editor.execute([InsertNewlineAtCaretRequest()]);
+  context.editor.execute([
+    InsertNewlineAtCaretRequest(),
+  ]);
 }
 
 void deleteWordBackward(SuperEditorContext context) {
