+++ super_editor_markdown/pubspec.yaml
@@ -30,8 +30,8 @@ dependency_overrides:
     path: ../super_editor
 #  super_text_layout:
 #    path: ../super_text_layout
-#  attributed_text:
-#    path: ../attributed_text
+  attributed_text:
+    path: ../attributed_text
 
 dev_dependencies:
   flutter_lints: ^2.0.1
