+++ super_editor_quill/pubspec.yaml
@@ -37,8 +37,8 @@ dependency_overrides:
     path: ../super_editor
 #  super_text_layout:
 #    path: ../super_text_layout
-#  attributed_text:
-#    path: ../attributed_text
+  attributed_text:
+    path: ../attributed_text
 
 dev_dependencies:
   # Note: some testing dependencies are listed as regular dependencies
